+++ docs/src/components/mapping-function-card.astro
@@ -1,6 +1,6 @@
 ---
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { mappingUrl, type Library } from "@/lib/mappings";
+import { mappingUrl, type Library } from "@/lib/libraries";
 import { ArrowRightIcon, ExternalLinkIcon } from "@radix-ui/react-icons";
 import { getEntry, render } from "astro:content";
 import invariant from "tiny-invariant";
