+++ docs/src/lib/mappings.ts
@@ -14,9 +14,6 @@ import type { CategorizedFunctions } from "./navbar-entries";
 
 const COLLECTION = "mapping";
 
-export const ALL_LIBRARIES = ["lodash", "ramda", "just"] as const;
-export type Library = (typeof ALL_LIBRARIES)[number];
-
 export async function getMappingEntries(
   library: string,
 ): Promise<CategorizedFunctions> {
@@ -41,17 +38,4 @@ export async function getMappingEntries(
     ),
     entries(),
   );
-}
-
-export function mappingUrl(library: Library, name: string): string {
-  switch (library) {
-    case "lodash":
-      return `https://lodash.com/docs/4.17.15#${name}`;
-
-    case "ramda":
-      return `https://ramdajs.com/docs/#${name}`;
-
-    case "just":
-      return `https://anguscroll.com/just/just-${name}`;
-  }
 }
