+++ docs/src/pages/migrate/[library].astro
@@ -2,16 +2,15 @@
 import DocsSection from "@/components/docs-section.astro";
 import MappingFunctionCard from "@/components/mapping-function-card.astro";
 import DocsPage from "@/layouts/docs-page.astro";
+import { ALL_LIBRARIES } from "@/lib/libraries";
 import { getMappingEntries } from "@/lib/mappings";
 import type { GetStaticPaths, InferGetStaticParamsType } from "astro";
 import { sortBy } from "remeda";
 
 export const getStaticPaths = (async () =>
-  [
-    { params: { library: "lodash" } },
-    { params: { library: "ramda" } },
-    { params: { library: "just" } },
-  ] as const) satisfies GetStaticPaths;
+  ALL_LIBRARIES.map((library) => ({
+    params: { library },
+  }))) satisfies GetStaticPaths;
 
 const { library } = Astro.params as InferGetStaticParamsType<
   typeof getStaticPaths
